Fixed float values handled in spinners showing angles
git-svn-id: https://svn.code.sf.net/p/sweethome3d/code/trunk/SweetHome3D@6137 03e323e1-6c6c-46a5-9f71-c4ce5d8c6204

+++ src/com/eteks/sweethome3d/swing/TextureChoiceComponent.java
@@ -428,8 +428,8 @@ public class TextureChoiceComponent extends JButton implements TextureChoiceView
       HomeTexture texture = controller.getTexture();
       setPreviewTexture(texture);
       if (texture instanceof HomeTexture) {
-        this.angleSpinner.setValue(Math.toDegrees(texture.getAngle()));
-        this.scaleSpinner.setValue(texture.getScale() * 100);
+        this.angleSpinner.setValue((float)Math.toDegrees(texture.getAngle()));
+        this.scaleSpinner.setValue(texture.getScale() * 100f);
       }
       Insets insets = border.getBorderInsets(this.texturePreviewComponent);
       this.texturePreviewComponent.setPreferredSize(
